Moved dict_id to IPC-specific IO
