+++ src/io/ipc/read/array/list.rs
@@ -1,25 +1,27 @@
-use std::collections::{HashMap, VecDeque};
+use std::collections::VecDeque;
 use std::convert::TryInto;
 use std::io::{Read, Seek};
-use std::sync::Arc;
 
 use arrow_format::ipc;
 
-use crate::array::{Array, ListArray, Offset};
+use crate::array::{ListArray, Offset};
 use crate::buffer::Buffer;
 use crate::datatypes::DataType;
 use crate::error::Result;
 
+use super::super::super::IpcField;
 use super::super::deserialize::{read, skip, Node};
 use super::super::read_basic::*;
+use super::super::Dictionaries;
 
 #[allow(clippy::too_many_arguments)]
 pub fn read_list<O: Offset, R: Read + Seek>(
     field_nodes: &mut VecDeque<Node>,
     data_type: DataType,
+    ipc_field: &IpcField,
     buffers: &mut VecDeque<&ipc::Schema::Buffer>,
     reader: &mut R,
-    dictionaries: &HashMap<usize, Arc<dyn Array>>,
+    dictionaries: &Dictionaries,
     block_offset: u64,
     is_little_endian: bool,
     compression: Option<ipc::Message::BodyCompression>,
@@ -55,6 +57,7 @@ where
     let values = read(
         field_nodes,
         field,
+        &ipc_field.fields[0],
         buffers,
         reader,
         dictionaries,
