+++ src/io/ipc/read/array/struct_.rs
@@ -1,23 +1,25 @@
-use std::collections::{HashMap, VecDeque};
+use std::collections::VecDeque;
 use std::io::{Read, Seek};
-use std::sync::Arc;
 
 use arrow_format::ipc;
 
-use crate::array::{Array, StructArray};
+use crate::array::StructArray;
 use crate::datatypes::DataType;
 use crate::error::Result;
 
+use super::super::super::IpcField;
 use super::super::deserialize::{read, skip, Node};
 use super::super::read_basic::*;
+use super::super::Dictionaries;
 
 #[allow(clippy::too_many_arguments)]
 pub fn read_struct<R: Read + Seek>(
     field_nodes: &mut VecDeque<Node>,
     data_type: DataType,
+    ipc_field: &IpcField,
     buffers: &mut VecDeque<&ipc::Schema::Buffer>,
     reader: &mut R,
-    dictionaries: &HashMap<usize, Arc<dyn Array>>,
+    dictionaries: &Dictionaries,
     block_offset: u64,
     is_little_endian: bool,
     compression: Option<ipc::Message::BodyCompression>,
@@ -38,10 +40,12 @@ pub fn read_struct<R: Read + Seek>(
 
     let values = fields
         .iter()
-        .map(|field| {
+        .zip(ipc_field.fields.iter())
+        .map(|(field, ipc_field)| {
             read(
                 field_nodes,
                 field,
+                ipc_field,
                 buffers,
                 reader,
                 dictionaries,
