+++ src/io/ipc/write/mod.rs
@@ -18,3 +18,49 @@ mod common_async;
 #[cfg(feature = "io_ipc_write_async")]
 #[cfg_attr(docsrs, doc(cfg(feature = "io_ipc_write_async")))]
 pub mod stream_async;
+
+use crate::datatypes::{DataType, Field};
+
+use super::IpcField;
+
+fn default_ipc_field(data_type: &DataType, current_id: &mut i64) -> IpcField {
+    use crate::datatypes::DataType::*;
+    match data_type.to_logical_type() {
+        // single child => recurse
+        Map(inner, ..) | FixedSizeList(inner, _) | LargeList(inner) | List(inner) => IpcField {
+            fields: vec![default_ipc_field(inner.data_type(), current_id)],
+            dictionary_id: None,
+        },
+        // multiple children => recurse
+        Union(fields, ..) | Struct(fields) => IpcField {
+            fields: fields
+                .iter()
+                .map(|f| default_ipc_field(f.data_type(), current_id))
+                .collect(),
+            dictionary_id: None,
+        },
+        // dictionary => current_id
+        Dictionary(_, data_type, _) => {
+            let dictionary_id = Some(*current_id);
+            *current_id += 1;
+            IpcField {
+                fields: vec![default_ipc_field(data_type, current_id)],
+                dictionary_id,
+            }
+        }
+        // no children => do nothing
+        _ => IpcField {
+            fields: vec![],
+            dictionary_id: None,
+        },
+    }
+}
+
+/// Assigns every dictionary field a unique ID
+pub fn default_ipc_fields(fields: &[Field]) -> Vec<IpcField> {
+    let mut dictionary_id = 0i64;
+    fields
+        .iter()
+        .map(|field| default_ipc_field(field.data_type().to_logical_type(), &mut dictionary_id))
+        .collect()
+}
