+++ src/io/ipc/write/stream.rs
@@ -1,20 +1,3 @@
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
 //! Arrow IPC File and Stream Writers
 //!
 //! The `FileWriter` and `StreamWriter` have similar interfaces,
@@ -22,7 +5,8 @@
 
 use std::io::Write;
 
-use super::common::{encoded_batch, DictionaryTracker, EncodedData, WriteOptions};
+use super::super::IpcField;
+use super::common::{encode_columns, DictionaryTracker, EncodedData, WriteOptions};
 use super::common_sync::{write_continuation, write_message};
 use super::schema_to_bytes;
 
@@ -48,24 +32,28 @@ pub struct StreamWriter<W: Write> {
 }
 
 impl<W: Write> StreamWriter<W> {
-    /// Try create a new writer, with the schema written as part of the header
-    pub fn try_new(mut writer: W, schema: &Schema, write_options: WriteOptions) -> Result<Self> {
-        // write the schema, set the written bytes to the schema
-        let encoded_message = EncodedData {
-            ipc_message: schema_to_bytes(schema),
-            arrow_data: vec![],
-        };
-        write_message(&mut writer, encoded_message)?;
-        Ok(Self {
+    /// Creates a new [`StreamWriter`]
+    pub fn new(writer: W, write_options: WriteOptions) -> Self {
+        Self {
             writer,
             write_options,
             finished: false,
             dictionary_tracker: DictionaryTracker::new(false),
-        })
+        }
     }
 
-    /// Write a record batch to the stream
-    pub fn write(&mut self, batch: &RecordBatch) -> Result<()> {
+    /// Starts the stream
+    pub fn start(&mut self, schema: &Schema, ipc_fields: &[IpcField]) -> Result<()> {
+        let encoded_message = EncodedData {
+            ipc_message: schema_to_bytes(schema, ipc_fields),
+            arrow_data: vec![],
+        };
+        write_message(&mut self.writer, encoded_message)?;
+        Ok(())
+    }
+
+    /// Writes [`RecordBatch`] to the stream
+    pub fn write(&mut self, batch: &RecordBatch, fields: &[IpcField]) -> Result<()> {
         if self.finished {
             return Err(ArrowError::Io(std::io::Error::new(
                 std::io::ErrorKind::UnexpectedEof,
@@ -73,8 +61,13 @@ impl<W: Write> StreamWriter<W> {
             )));
         }
 
-        let (encoded_dictionaries, encoded_message) =
-            encoded_batch(batch, &mut self.dictionary_tracker, &self.write_options)?;
+        let columns = batch.clone().into();
+        let (encoded_dictionaries, encoded_message) = encode_columns(
+            &columns,
+            fields,
+            &mut self.dictionary_tracker,
+            &self.write_options,
+        )?;
 
         for encoded_dictionary in encoded_dictionaries {
             write_message(&mut self.writer, encoded_dictionary)?;
