+++ tests/it/io/ipc/common.rs
@@ -1,18 +1,18 @@
-use std::{collections::HashMap, convert::TryFrom, fs::File, io::Read};
+use std::{collections::HashMap, fs::File, io::Read};
 
 use arrow2::{
-    datatypes::Schema,
-    error::Result,
-    io::ipc::read::read_stream_metadata,
-    io::ipc::read::StreamReader,
-    io::json_integration::{to_record_batch, ArrowJson},
-    record_batch::RecordBatch,
+    datatypes::Schema, error::Result, io::ipc::read::read_stream_metadata,
+    io::ipc::read::StreamReader, io::ipc::IpcField, io::json_integration::read,
+    io::json_integration::ArrowJson, record_batch::RecordBatch,
 };
 
 use flate2::read::GzDecoder;
 
 /// Read gzipped JSON file
-pub fn read_gzip_json(version: &str, file_name: &str) -> Result<(Schema, Vec<RecordBatch>)> {
+pub fn read_gzip_json(
+    version: &str,
+    file_name: &str,
+) -> Result<(Schema, Vec<IpcField>, Vec<RecordBatch>)> {
     let testdata = crate::test_util::arrow_test_data();
     let file = File::open(format!(
         "{}/arrow-ipc-stream/integration/{}/{}.json.gz",
@@ -27,7 +27,7 @@ pub fn read_gzip_json(version: &str, file_name: &str) -> Result<(Schema, Vec<Rec
 
     let schema = serde_json::to_value(arrow_json.schema).unwrap();
 
-    let schema = Schema::try_from(&schema)?;
+    let (schema, ipc_fields) = read::deserialize_schema(&schema)?;
 
     // read dictionaries
     let mut dictionaries = HashMap::new();
@@ -41,13 +41,16 @@ pub fn read_gzip_json(version: &str, file_name: &str) -> Result<(Schema, Vec<Rec
     let batches = arrow_json
         .batches
         .iter()
-        .map(|batch| to_record_batch(&schema, batch, &dictionaries))
+        .map(|batch| read::to_record_batch(&schema, &ipc_fields, batch, &dictionaries))
         .collect::<Result<Vec<_>>>()?;
 
-    Ok((schema, batches))
+    Ok((schema, ipc_fields, batches))
 }
 
-pub fn read_arrow_stream(version: &str, file_name: &str) -> (Schema, Vec<RecordBatch>) {
+pub fn read_arrow_stream(
+    version: &str,
+    file_name: &str,
+) -> (Schema, Vec<IpcField>, Vec<RecordBatch>) {
     let testdata = crate::test_util::arrow_test_data();
     let mut file = File::open(format!(
         "{}/arrow-ipc-stream/integration/{}/{}.stream",
@@ -58,10 +61,12 @@ pub fn read_arrow_stream(version: &str, file_name: &str) -> (Schema, Vec<RecordB
     let metadata = read_stream_metadata(&mut file).unwrap();
     let reader = StreamReader::new(file, metadata);
 
-    let schema = reader.schema();
+    let schema = reader.metadata().schema.as_ref().clone();
+    let ipc_fields = reader.metadata().ipc_schema.fields.clone();
 
     (
-        schema.as_ref().clone(),
+        schema,
+        ipc_fields,
         reader
             .map(|x| x.map(|x| x.unwrap()))
             .collect::<Result<_>>()
