+++ tests/it/io/ipc/write/file.rs
@@ -1,34 +1,40 @@
 use std::io::Cursor;
 
 use arrow2::array::*;
+use arrow2::datatypes::Schema;
 use arrow2::error::Result;
 use arrow2::io::ipc::read::{read_file_metadata, FileReader};
-use arrow2::io::ipc::write::*;
+use arrow2::io::ipc::{write::*, IpcField};
 use arrow2::record_batch::RecordBatch;
 
 use crate::io::ipc::common::read_gzip_json;
 
-fn round_trip(batch: RecordBatch) -> Result<()> {
-    let result = Vec::<u8>::new();
+fn write_(
+    batches: &[RecordBatch],
+    schema: &Schema,
+    ipc_fields: Option<Vec<IpcField>>,
+    compression: Option<Compression>,
+) -> Result<Vec<u8>> {
+    let result = vec![];
+    let options = WriteOptions { compression };
+    let mut writer = FileWriter::try_new(result, schema, ipc_fields.clone(), options)?;
+    for batch in batches {
+        writer.write(batch, ipc_fields.as_ref().map(|x| x.as_ref()))?;
+    }
+    writer.finish()?;
+    Ok(writer.into_inner())
+}
 
-    // write IPC version 5
-    let written_result = {
-        let options = WriteOptions {
-            compression: Some(Compression::LZ4),
-        };
-        let mut writer = FileWriter::try_new(result, batch.schema(), options)?;
-        writer.write(&batch)?;
-        writer.finish()?;
-        writer.into_inner()
-    };
-    let mut reader = Cursor::new(written_result);
+fn round_trip(batch: RecordBatch, ipc_fields: Option<Vec<IpcField>>) -> Result<()> {
+    let (expected_schema, expected_batches) = (batch.schema().clone(), vec![batch.clone()]);
+
+    let schema = batch.schema().clone();
+    let result = write_(&[batch], &schema, ipc_fields, Some(Compression::ZSTD))?;
+    let mut reader = Cursor::new(result);
     let metadata = read_file_metadata(&mut reader)?;
     let schema = metadata.schema().clone();
 
     let reader = FileReader::new(reader, metadata, None);
-
-    // read expected JSON output
-    let (expected_schema, expected_batches) = (batch.schema().clone(), vec![batch]);
 
     assert_eq!(schema.as_ref(), expected_schema.as_ref());
 
@@ -39,9 +45,7 @@ fn round_trip(batch: RecordBatch) -> Result<()> {
 }
 
 fn test_file(version: &str, file_name: &str, compressed: bool) -> Result<()> {
-    let (schema, batches) = read_gzip_json(version, file_name)?;
-
-    let result = Vec::<u8>::new();
+    let (schema, ipc_fields, batches) = read_gzip_json(version, file_name)?;
 
     let compression = if compressed {
         Some(Compression::ZSTD)
@@ -49,26 +53,20 @@ fn test_file(version: &str, file_name: &str, compressed: bool) -> Result<()> {
         None
     };
 
-    // write IPC version 5
-    let written_result = {
-        let options = WriteOptions { compression };
-        let mut writer = FileWriter::try_new(result, &schema, options)?;
-        for batch in batches {
-            writer.write(&batch)?;
-        }
-        writer.finish()?;
-        writer.into_inner()
-    };
-    let mut reader = Cursor::new(written_result);
+    let result = write_(&batches, &schema, Some(ipc_fields), compression)?;
+    let mut reader = Cursor::new(result);
     let metadata = read_file_metadata(&mut reader)?;
     let schema = metadata.schema().clone();
+    let ipc_fields = metadata.ipc_schema.fields.clone();
 
     let reader = FileReader::new(reader, metadata, None);
 
     // read expected JSON output
-    let (expected_schema, expected_batches) = read_gzip_json(version, file_name)?;
+    let (expected_schema, expected_ipc_fields, expected_batches) =
+        read_gzip_json(version, file_name)?;
 
     assert_eq!(schema.as_ref(), &expected_schema);
+    assert_eq!(ipc_fields, expected_ipc_fields);
 
     let batches = reader.collect::<Result<Vec<_>>>()?;
 
@@ -332,7 +330,7 @@ fn write_boolean() -> Result<()> {
         Some(true),
     ])) as Arc<dyn Array>;
     let batch = RecordBatch::try_from_iter(vec![("a", array)])?;
-    round_trip(batch)
+    round_trip(batch, None)
 }
 
 #[test]
@@ -341,7 +339,7 @@ fn write_sliced_utf8() -> Result<()> {
     use std::sync::Arc;
     let array = Arc::new(Utf8Array::<i32>::from_slice(["aa", "bb"]).slice(1, 1)) as Arc<dyn Array>;
     let batch = RecordBatch::try_from_iter(vec![("a", array)])?;
-    round_trip(batch)
+    round_trip(batch, None)
 }
 
 #[test]
@@ -357,5 +355,5 @@ fn write_sliced_list() -> Result<()> {
     array.try_extend(data).unwrap();
     let array = array.into_arc().slice(1, 2).into();
     let batch = RecordBatch::try_from_iter(vec![("a", array)]).unwrap();
-    round_trip(batch)
+    round_trip(batch, None)
 }
